requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
cffi==1.15.1
cfgv==3.3.1
cryptography==38.0.4
distlib==0.3.6
Django==4.1.4
django-ninja==0.20.0
django-redis==5.2.0
filelock==3.9.0
identify==2.5.11
install==1.3.5
nodeenv==1.7.0
Pillow==9.4.0
platformdirs==2.6.2
pre-commit==2.21.0
py3Fdfs==2.2.0
pycparser==2.21
pydantic==1.10.4
PyJWT==2.6.0
PyMySQL==1.0.2
PyYAML==6.0
redis==4.4.0
sqlparse==0.4.3
typing_extensions==4.4.0
virtualenv==20.17.1
